# Make replaceShortcodeTokens modify the input slice