@bep 2015-11-13
/cc @spf13 

I have tested this green with all of my sites -- and I don't see any downsides with it ... But a second set of eyes is always good.