@bep 2015-11-17
I have to look a little close on this method's many uses -- will do tonight.