@bep 2015-11-17
Merged in b08d9f26ae00b6c55f615ad6b8fc36608e684204